Import everythng after we do the command
